git-svn-id: http://www.ifountain.org/repos/os/RapidModules@544 a08bcbf9-5f6c-4d33-af0a-e494bc2cacca

+++ RapidCMDB/rs.sh
@@ -16,5 +16,10 @@ if [ "$RS_HOME" == "" ]; then
 fi
 
 export GRAILS_HOME=$RS_HOME
-rm -r $RS_HOME/temp
-. "%GRAILS_HOME%\bin\startGrails" "%GRAILS_HOME%" org.codehaus.groovy.grails.cli.GrailsScriptRunner  %* -Dserver.port=12222 -Dgrails.work.dir=$RS_HOME/temp
+
+if [ -f "$RS_HOME/temp" ]
+then 
+	rm -r $RS_HOME/temp
+fi
+. $GRAILS_HOME/bin/startGrails 
+startGrails org.codehaus.groovy.grails.cli.GrailsScriptRunner  "$@" -Dserver.port=12222 -Dgrails.work.dir=$RS_HOME/temp
